fix: move packets in correct directory
